+++ playhouse/apsw_ext.py
@@ -101,6 +101,9 @@ class APSWDatabase(SqliteDatabase):
         self.cursor().execute('begin %s;' % lock_type)
 
     def commit(self):
+        if self.is_closed():
+            raise InterfaceError('Cannot commit, database connection not '
+                                 'open.')
         with __exception_wrapper__:
             curs = self.cursor()
             if curs.connection.getautocommit():
@@ -109,6 +112,9 @@ class APSWDatabase(SqliteDatabase):
         return True
 
     def rollback(self):
+        if self.is_closed():
+            raise InterfaceError('Cannot rollback, database connection not '
+                                 'open.')
         with __exception_wrapper__:
             curs = self.cursor()
             if curs.connection.getautocommit():